requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
pydantic[email]==2.7.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
celery[redis]==5.4.0
redis==5.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.131
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
openai==1.35.7
slowapi==0.1.9
structlog==24.2.0
tenacity==8.3.0
python-magic==0.4.27
httpx==0.27.0
